limited ephemeral node support
each client gets a /eph symlink to /ctl/node/NODEID/client/IP-PORT
which is the ephemeral node for that client. this node automatically
gets deleted on client disconnect. thus, other clients can simply glob
over /ctl/node/*/client/* to find the current list of active clients.
additionally, the client can set the value of this ephemeral node by
calling SET('/eph', 'who am i', rev).

this feature is used to determine the list of running processes (which
would keep a persistent connection to doozer) on each node in the
cluster, and consequently have the process manager on those nodes
auto-start crashed processes.

known issue: the webui will freeze sometimes, but this is a known bug
even before this change which only seemed to have triggered it.